requirements.txt
asgiref==3.2.3
bcrypt==3.1.7
cffi==1.14.0
cryptography==2.8
Django==3.0.3
django-environ==0.4.5
django-extensions==2.2.8
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
fabric==2.5.0
fabtools==0.20.0
gunicorn==20.0.4
invoke==1.4.1
paramiko==2.7.1
Pillow==7.0.0
psycopg2-binary==2.8.4
pycparser==2.19
PyJWT==1.7.1
PyNaCl==1.3.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
